Forenklet patchning av reell arbeidssøker
